refactor: to domain folder structure
