+++ src/domain/application/use-cases/authenticate.spec.ts
