+++ src/domain/application/use-cases/get-user-profile.spec.ts
