+++ src/domain/application/use-cases/register.spec.ts
